+++ minecraft/paper/src/main/java/broccolai/corn/paper/item/special/SkullBuilder.java
@@ -1,5 +1,6 @@
-package broccolai.corn.paper.item;
+package broccolai.corn.paper.item.special;
 
+import broccolai.corn.paper.item.AbstractPaperItemBuilder;
 import com.destroystokyo.paper.profile.PlayerProfile;
 import com.destroystokyo.paper.profile.ProfileProperty;
 import org.bukkit.Material;
