+++ minecraft/paper/src/main/java/broccolai/corn/paper/item/special/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Builders specific to certain {@link org.bukkit.inventory.meta.ItemMeta}s.
+ */
+package broccolai.corn.paper.item.special;
